+++ docs/CHANGELOG.md
@@ -14,14 +14,14 @@ Entries inside each section should be ordered by type:
 
 ## Catalog, Lambdas
 !-->
-# unreleased
+# 5.3.1 - 2023-05-02
 ## Python API
-
-## CLI
+* [Fixed] `Package.verify()` now raises exception if unsupported hash type is encountered ([#3401](https://github.com/quiltdata/quilt/pull/3401))
 
 ## Catalog, Lambdas
 * [Fixed] Fix file URLs in embed ([#3419](https://github.com/quiltdata/quilt/pull/3419))
 * [Added] Make package dialog controllable via URL and create Dropzone-only package dialog ([#3405](https://github.com/quiltdata/quilt/pull/3405))
+* [Changed] Increased available file size to render Markdown preview to 3MiB ([#3427](https://github.com/quiltdata/quilt/pull/3427))
 
 # 5.3.0 - 2023-04-11
 ## Python API